+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* Fix `S3#multi_delete` and `S3#clear!` not using the prefix (janko-m)
+
 * Add ability to pass presign options to storages in the direct_upload plugin (janko-m)
 
 * Remove `Shrine.io!` because it was actually meant to be only for internal use (janko-m)